package: Always normalize Package.package_name
The .package_name field is not used for anything important, but
resolvers have varied as to whether they normalize this member or not.

As of now, only IdentityMapping retained the original spelling of
.package_name, as the other resolver all depend on accumulate_mappings()
which normalize the resulting .package_names.

This settles the debate by having Package itself ensure that the
.package_name is always normalized.