TemporaryPipInstallResolver: Handle 'pip install' errors gracefully
Until now, we have not handled the CalledProcessError that is raised
when `pip install` fails to install the given requirements in
TemporaryPipInstallResolver.temp_installed_requirements().

This commit changes that by, first, disabling the CalledProcessError
and instead logging a warning message, and second, re-running
`pip install` on each individual requirement in order. The effect of
this will be to install the subset of the given requirements that we
_can_ install, leaving only the problematic requirements uninstalled.

This enables TemporaryPipInstallResolver to do a "best effort"
installation of dependencies: Inside the .temp_installed_requirements()
context, the LocalPackageResolver will be able to resolve those
dependencies that were successfully installed, and the dependencies we
failed to install will be handed off to the next level of the resolver
stack.

As of writing this is always the IdentityMapping, but this is about to
go away, and instead a proper UnresolvedDependenciesError will be
raised. Together with the warning messages from the failed `pip install`
commands, this should inform the user of exactly which
packages/dependencies we failed to install, and why.

On the tests/ side, move our existing TemporaryPipInstallResolver test
into a new test module: test_install_deps, and add two more tests:

 - to test that we no longer propagate a CalledProcessError when `pip
   install` fails (but instead log a warning message), and
 - to test that we are able to install one dependency with `pip install`
   even if another dependency fails.

This should fix issue #322 (Handle `pip install` errors in
TemporaryPipInstallResolver)