Make TemporaryPipInstallResolver independent of pytest
The last changes defined the cache in TemporaryPipInstallResolver with
Pytest's cache. We introduce a new class member `cached_venv`, to
TemporaryPipInstallResolver. This is only used by `test_resolver()`
where it's set to a Pytest's cache directory, thus removing the direct
dependency of TemporaryPipInstallResolver on Pytest.

Furthermore, the new TemporaryPipInstallResolver context manager
introduced in the recent changes reuses code for intsalling requirements
and managing the venv from `temp_installed_requirements()`.
This commit splits out this duplicate code into a context manager,
`installed_requirements()`, that manages the venv creation (if needed)
and installs packages.
The existing `temp_installed_requirements()` context manager will then
only concern itself with creating the temporary directory (and delegate
to installed_requirements() for the rest).

As this replaces the initialization logic in TemporaryPipInstallResolver
with a simple class member that defines the cached_venv, it allows us to
avoid depending on pytest-mock for patching the class.

Co-authored-by: Johan Herland <johan.herland@tweag.io>