shell.nix: Add support for installing wheels with compiled binaries
We want to start using ruff in FawltyDeps, but this does not work out of
the box in our Nix dev shell: The ruff wheel installed by Poetry
contains a pre-compiled version of ruff, i.e. an ELF executable that
expects to find an ELF interpreter at /lib64/ld-linux-x86-64.so.2. This
breaks the hermeticity in Nix, and on certain systems (NixOS in
particular) it fails to work at all, since that interpreter path does
not exist.

This commit teaches our Nix dev shell to:
 - provide all libraries necessary to support binary wheels following
   the manylinux2014 standard (see: https://peps.python.org/pep-0599/)
 - automatically patch any ELF binaries installed into the Poetry
   virtualenv so as to rely on the Nix-provided ELF interpreter and
   libraries.

This re-establishes hermeticity for pre-compiled wheels, even though we
still rely on Poetry to install project dependencies.