main: Add Analysis.sources to JSON output
Increase transparency and debug-friendliness by including our Source
objects in the --json output.

In order to more easily differentiate the various types of Source
objects in this output, we want to include the Source subclass name in
its JSON representation. This is accomplished by making a proper Source
base class that automatically injects a .source_type member which
records the subclass. This is then included in the JSON representation.

Without this, a CodeSource, DepsSource, and PyEnvSource would be
represented like this, and differentiating them would have to be done by
looking at incidental members:

        "sources": [
            {
                "path": "/some/path/code.py",
                "base_dir": "/some/path",
            },
            {
                "path": "/some/path/requirements.txt",
                "parser_choice": "requirements.txt",
            },
	    {
	        "path": "/some/path/.venv",
	    },
        ]

Instead, we make the entries self-documenting, by including the subclass
name:

        "sources": [
            {
                "source_type": "CodeSource",
                "path": "/some/path/code.py",
                "base_dir": "/some/path",
            },
            {
                "source_type": "DepsSource",
                "path": "/some/path/requirements.txt",
                "parser_choice": "requirements.txt",
            },
	    {
	        "source_type": "PyEnvSource",
		"path": "/some/path/.venv",
	    },
        ],