fake_venv: Allow customization of venv location and Python version
This allows future users of the fake_venv fixture to easily choose
_where_ the fake venv is created, and what Python version is used for
the lib dir where the _fake_ packages are installed.

Note that the venv itself is created with `venv.create()` and is hance a
"real" virtualenv whose bin/python is always the current Python version.
What makes it "fake" and appear to be for a different Python version, is
the creation of files under lib/pythonX.Y/site-packages, where X.Y
reflects the given 'py_version'.