Add --list-sources CLI argument
This allows FawltyDeps to stop early and print the sources that would be
used for finding imports and declared dependencies in the project.

This can be used to help find the appropriate options to pass to
FawltyDeps in order to have it analyze the desired files.

The (default) --summary view lists all file paths that are used as
sources. The --detailed view also categorizes sources by type, and adds
extra information relevant to the type (for CodeSource, which base path
is used to recognize 1st-party imports, and for DepsSource, which parser
is used to parse declared dependencies).