shell.nix: Update our development environment
Before we can start thinking about supporting the upcoming Python v3.12
release, we must upgrade our Nix-based dev shell to a version of Nixpkgs
that supports/contains python312. However, there is no version of
nixpkgs that supports _both_ Python v3.7 (which recently reached its
end-of-life), and Python v3.12. This can be solved, however, by pulling
in python37 from an older version of nixpkgs.

This commit upgrades our shell.nix to a newer version of nixpkgs (while
still pulling python37 from an older version). The new version of
nixpkgs has also changed how it provides Poetry: Instead of pulling it
as a Python package with _one_ of the Python versions we use, it is now
consumed as a top-level package. As a side effect, we now start using a
newer Poetry version:v1.4.2 [1]. The Python version that we use inside
the default Poetry shell is then determined by which python executable
we pass to the `poetry env use` command. For now, this remains the same
Python v3.10 that we have been using for a while.

[1]: The new Poetry version considerably reformats the poetry.lock file,
but there should be no semantic changes to this file: the current
version was generated by running `poetry lock --no-update` after
entering the new nix-shell.