cli_parser.py: Rephrase the help text of basepaths/--code/--deps/--pyenv
The --code/--deps/--pyenv options was documented as defaulting to the
current directory (which is only true when basepaths is not used), and
basepaths was NOT documented as defaulting to the current directory.

The correct way to document this (which reflects the actual logic in the
code) is that --code/--deps/--pyenv all default to basepaths when not
given, and that basepaths default to the current directory when not
given.

With this change, a casual reader might have to follow the reference
from --code/--deps/--pyenv to basepaths to discover that the _ultimate_
default for these options is indeed the current directory. But at least
the relationships involving basepaths are now _correctly_ reflected in
out documentation.

Suggested-by: Maria Knorps <maria.knorps@tweag.io>