test_settings: Prepare for including --pyenv in basepath-related tests
We have some test cases that verify correct interaction between
--code/--deps and basepath. We're about to add --pyenv to --code and
--deps as another option that interacts with basepath, and we'd like the
tests to cover this as well.

At this point the various combinations of passing or not passing these
options on the command line are becoming tedious to hardcode.

Instead, provide a `path_options` collection that list these options, and
a subsets() helper function that automatically enumerates the various
combinations of giving or not giving these options the we need in
test_path_option_overrides_base_path().