Settings: Add .exclude, paths to exclude during project traversal
The new Settings.exclude member is a set of strings, where each string
is a .gitignore pattern that is applied during project traversal. Paths
matching a pattern in Settings.exclude will not be traversed while
looking for code, deps, or pyenvs.[1]

The default value of Settings.exclude is {".*"}, that is a single
exclude pattern that matches hidden/dot paths. This replaces the
hardcoded exclusion of hidden/dot paths that was done previously.

In order to test Settings.exclude, this commit also adds support for
testing exclude patterns to test_traverse_project.

[1]: Minor exception: As detailed in a previous commit, a Python
environment located exactly at an excluded directory will still be
found, but we won't traverse into an excluded directory to find Python
environments.