dir_traversal: Rename .ignore() to .skip_dir()
We're about to introduce more functionality for ignoring files/dirs in a
directory structure via .gitignore-style patterns.

However, we're already using the term "ignore" to mean something
slightly different in DirectoryTraversal: The .ignore() method is used
to tell the traversal that some directory should _not_ be traversed.

Rename this method to .skip_dir() (and adjust members and comments
accordingly). This is more accurate to what this method actually does,
and it frees up the "ignore" term to be used in future commits.