DirectoryTraversal: Initial support for exclude patterns
Use the gitignore_parser package to parse .gitignore-style patterns
passed to the new DirectoryTraversal.exclude() method. The result of
this is a list of IgnoreRules that can be tested against a given path to
determine whether that path should be excluded from traversal.

Next, for each directory being traversed, calculate the subsets of
subdirs and filenames that should be excluded, and subtract these from
the sets that are passed to our caller. Also make sure that excluded
subdirs are passed to self.skip_dir(), so that these are not traversed.

Add a comprehensive test suite that tests all the .gitignore patterns
documented at https://git-scm.com/docs/gitignore#_pattern_format.