Upgrade Pylint for Python versions >=v3.8
Our current Pylint version crashes when run with Python v3.12, and we
need to upgrade to Pylint v3 in order to fix this. However, Pylint v3
removes support for Python v3.7. Hence, Upgrade Pylint to v3 for Python
>=v3.8, and remain on the old Pylint version for Python v3.7.

In addition to the pyproject.toml/poetry.lock upgrade itself, there are
a few new warnings that need silencing in order for our code to pass
with the new Pylint version.