extract_imports.parse_code(): Accept code as bytes object
parse_code() takes Python code and passes it directly on to
ast.parse(), which again passed it to the compile() built-in function.
compile() handles code passed as either str objects, or bytes objects,
so parse_code() can handle this as well, we just need to update the type
signature and documentation.

The difference between the two cases, AFAICS, is that in the bytes case
the decoding of the source text happens inside compile() (and it can
therefore auto-detect the appropriate encoding as described in [1]).
In the str case, however, compile() must _assume_ that the source text
has already been decoded with the appropriate encoding.

Ensuring that our callers meet this assumption for str objects will be
done in the following commits.

Add a test to demonstrate that passing legacy-encoded bytes directly to
parse_code() is handled correctly.

[1]: https://docs.python.org/3/reference/lexical_analysis.html#encoding-declarations