traverse_project: Use new exclude feature to exclude "hidden" paths
Instead of manually skipping dot dirs at each traversal step, use the
corresponding ".*" gitignore pattern to exclude dot dirs (and files).

In practice, this should not cause any changes for our users, but there
some complexity worth mentioning:

Before this commit, we skipped traversing into subdirs whose name
starts with ".", but these entries were still present in the
step.subdirs and step.files. This is no longer the case, and as a
consequence this commit introduces one change, and one non-change
(that still warrants explanation):

1. Unlike before, we will no longer read code or deps from files that
   start with ".", by default. I would argue that this is _more_ in line
   with our intention of ignoring dot/hidden paths. To read code or deps
   from a dot file, it must now be passed to --code/--deps explicitly.

2. As before, we make an exception to the dot rule for Python
   environments, when the pyenv itself lives _directly_ in a dot path.
   For examples, we will find a Python environment inside ".venv", but
   we will not find one in ".venvs/some_venv".

   In this commit this is accomplished by looking for Python
   environments in the union of step.subdirs and step.excluded_subdirs,
   but since the traversal never descends into step.excluded_subdirs we
   won't find any Python environments _inside_ these.

This might seem over-complicated, but it's a compromise between these
factors:
- We want to exclude code/deps in dot files (and inside dot dirs).
- We want to recognize Python environments that are commonly stored in
  dot dirs (like .venv/), but less commonly stored _inside_ dot dirs
  (like .venvs/foo).
- We don't want to present the user with a complex choice of options for
  ignoring code/deps vs. ignoring pyenvs.
- Walking the filesystem is expensive, we'd like to do it only _once_,
  i.e. we don't want separate traversals for code/deps vs. pyenvs.

In short, our default policy is that we do not find code/deps/pyenvs in
dot (aka. hidden) paths, with the exception that a pyenv located _at_
(not inside) a dot dir will still be recognized.

To further document this behavior, this commit adds a bunch of tests in
this area, to catch future regressions.