dir_traversal: Do not exclude an explicitly .add()ed path
If a path has been passed to DirectoryTraversal.add() (this
corresponds to it being listed explicitly in a config file or on the
command line), and the same path also happens to match an exclude
pattern, should the path be traversed or excluded?

It should be traversed. That is, we should prioritize an explicitly
listed paths over exclude patterns.

This did not happen before this commit: the matching exclude would cause
the subdir to be passed to DirectoryTraversal.skip_dir(), which would
disqualify it from later traversal.

Fix this, when we calculate the subdirs to be excluded, by checking if
the subdirs correspond to paths that remain to be traversed.