gitignore_parser: Tighten types in API: PathOrStr -> Path
Instead of accepting PathOrStr in most API functions, accept only Path
instead. A side-effect of this is that we must communicate whether a
path is a directory as a separate bool, since the trailing slash that
signifies a directory (a detail which was not well documneted anyway)
is not preserved in Path objects.