gitignore_parser: Fix dir_only rule matching files
This was implemented as a workaround in DirectoryTraversal._do_exclude()
where it relied on the fact that DirectoryTraversal always processes
parents before children, i.e. a dir_only rule exactly matching a
directory path (true positive) would always precede the same rule
prefix-matching a child path (which would have yielded a false positive,
but never occured due to the preceding match preventing the traversal of
any children).

Thus, the workaround worked well for the specific case of
DirectoryTraversal, but not for the general case of simply passing that
child path to gitignore_parser.

This commit fixes the issue properly, by making sure that the dir_only
bit only takes effect when the given path is not a dir, _and_ the
matched part of the path includes the very end of the path (i.e. we're
not looking at a child).