Always include current env in the package resolver stack
This fixes issue #392 by making sure that in the case where there are
only _irrelevant_ venvs located inside the project, we will still find
at least _one_ Python environment where we can have some hope of finding
relevant dependencies installed.

In fact, this is partially a return to our pre v0.7 behavior where we
would always (then: only) find packages in sys.path. Since v0.7 we
allowed --pyenv to override this, and since v0.12, we also started
searching for pyenvs under --pyenv or basepaths given on the command
line. The result of this is that by now, the current environment is only
consulted in the fallback where no other environment is possibly found.

The change here is minimal (due to the preparations in previous
commits), and the existing tests are also compatible with this change,
except for the self_test Nox target, which can now be simplified to no
longer pass a --pyenv option, due to fixing this exact issue.

Most of the diff here is simply updating the relevant documentation to
reflect and clarify our new policy.

In the future, in the unlikely event that including sys.path in the
resolver stack introduces any regression, we might consider adding a
user-visible option to disable this again.