Convince Poetry to install latest dependency versions on newer Pythons
Over-specify our main dependencies to get the latest versions of
dependencies installed on more recent Python versions. Without this,
the versions of these dependencies chosen by Poetry will be the ones
that are compatible with Python 3.7, even if the current Python version
is much newer.

We do this to improve our test coverage for dependency versions that are
more likely to be used when FawltyDeps are installed with pip (in which
case Poetry is not involved, and the looser version constraints in
pyproject.toml are followed).

See https://github.com/orgs/python-poetry/discussions/8857 and
https://github.com/jherland/poetry_overspecify_deps_demo for more
details.