gitignore_parser: Minor change in file parsing APIs
In preparation for adding gitignore file support in dir_traversal, we
make a small change to the return value of parse_gitignore() and
parse_gitignore_lines():

Instead of directly returning a callable that can be used to match lines
(i.e. a partially applied match_rules()), we rather return a sequence of
Rule objects directly, and leave it to the caller to organize these into
a list and pass that to match_rules(). This allows the caller to collect
rules across multiple calls to parse_gitignore(),
parse_gitignore_lines(), and Rule.from_pattern(), before they choose to
invoke match_rules().