Settings: Add .exclude_from, paths containing exclude patterns
This exposes the DirectoryTraversal.exclude_from() API to the Settings
layer in FawltyDeps, and allows configuration of which files we read
exclude patterns from.

What is NOT in this commit: A feature that we want to add is the ability
to automatically parse and respect exclude patterns from .gitignore
files (and friends, such as .ignore and .git/info/exclude). However,
sinve .gitignore files can be placed all over the project's source tree,
we cannot support this correctly with a statically configured path set
like .exclude_from. Instead, we need to make .gitignore detection part
of the traversal itself, and dynamically update the exclude patterns
accordingly. Ruff is a linter that IMHO does this correctly, by exposing
a boolean `respect-gitignore` directive (defaults to `true`) to control
this logic. (See https://docs.astral.sh/ruff/settings/#respect-gitignore
for more details.) The addition/implementation of a respect-gitignore
setting is deferred to a later PR.