test_dir_traversal: Add .skip_me member to each DirectoryTraversalVector
Instead of a complex pytest.mark.skipif() expression at the point where
we parametrize the test vectors, add a .skip_me member which allows
individual test vectors to customize under which conditions they are
skipped.

The .skip_me member is either None when the test vector should NOT be
skipped, or a string describing the reason for why this test vector
should be skipped. .skip_me defaults to `lambda: None`, meaning that the
test vector will never be skipped.

Also add an on_windows() helper to allow easy skipping of tests on
Windows.

Refactor the skipping of test vectors with symlinks to be skipped on
Windows using this new helper.

By itself, the value of this refactoring is questionable, but this
allows for other .skip_me conditions to be added to individual test
vectors in the future which would not be feasible to express in the
pytest.mark.skipif() expression.