dir_traversal: Fix DirId.from_path() with relative paths
We used @lru_cache to cache the results of DirId.from_path() (so as to
avoid unnecesary path.stat() calls). However, when passing a relative
path to DirId.from_path() this cache breaks down as soon as the current
working directory is changed: E.g. when calling DirId.from_path(".") we
would keep returning the same DirId without realizing that "." now
points to a _different_ directory.

Fix this by limiting the caching to absolute paths only, and then
replacing DirId.from_path() with a wrapper that makes all relative paths
absolute before calling the caching version.

This fixes a few of the currently-failing dir_traversal tests.