test_dir_traversal: Test relative paths
Add a test_DirectoryTraversal_w_rel_paths() test case that tests the
same DirectoryTraversalVectors as are already tested by
test_DirectoryTraversal_w_abs_paths(), but instead of passing an
absolute tmp_path into the traversal code, we instead run the test from
_inside_ tmp_path, and pass a simple "." relative path to the traversal
code.

This ends up testing tranversal scenarios that are closer to what users
typically run in the Real World: You run FawltyDpes on a project in (or
relative to) the current directory, you don't pass absolute paths to
FawltyDeps.

A number of the DirectoryTraversalVectors currently fail when run in
this new scenario. These failures will be fixed in the next commits.