test_gitignore_parser: Prepare for testing relative paths
Currently all these tests pass /some/dir/... absolute paths to the
gitignore_parser module. We want to verify that the gitignore_parser
works just as well with relative paths.

To prepare for this, we remove the /some/dir/ prefix from all test
vectors, and instead prepend the /some/dir/ prefix inside the test case
code. This does not change the existing test, but allows adding a new
test case later that will work with relative paths.