gitignore_parser: Fix dir-only patterns matching children of matched dir
Consider a dir-only patttern like "foo/". This should obviously match
"foo" directories like ./foo/ and ./some/other/foo/. But should it
match paths inside those directories, like ./foo/bar as well?

Until now, gitignore_parser has answered this questions with:

  "Yes, the ./foo/bar path is inside a directory ./foo/ that should be
  ignored, and therefore everything inside that directory should also be
  ignored."

This makes sense when you consider gitignore_parser in isolation: It
does not know or require any kind of context around what kind of paths
are passed, or in which order.

However, consider the following command line:

  fawltydeps --code=foo/ --exclude=foo/

On the surface, this instructs FD to look for code under foo/, but
also to exclude any foo/ directory from traversal. We have already
resolved this by declaring that --exclude shall lose a direct conflict
like this, and this is followed by our traversal code, we _do_ in fact
traverse into the foo/ directory.

However, since foo/ is a dir-only pattern, gitignore_parser will now
proceed to succesfully match any path that we traverse inside foo/
against this pattern, and the end result is the same as if we never
traversed into foo/: Nothing under foo/ is found/looked at by FD.

To solve this, we need to take some more context into account: The
directory traversal code always checks parent dirs before descending
into them, and when it has decided to traverse into a directory, we
must assume that any match against that directory has already been
considered. We cannot have an exclude pattern that matches this
directory only _also_ match paths inside that directory (and thus
in effect override the decision of traverse into it).

With this commit we will know only apply a dir-only pattern to the
_final_ path component, meaning that children of the matched directory
will no longer match the same pattern.

In effect we are _coupling_ gitignore_parser to the dir_traversal code,
here, and changing its answer to the above question as follows:

  "No, even though I think the ./foo directory should be ignored, if you
  ask me about the ./foo/bar path, I will trust that you have already
  considered my answer to ./foo before deciding to traverse into it. My
  answer will therefore reflect the bar path under foo/ specifically,
  and this bar path does not match the foo/ pattern."

This tighter coupling ends up actually simplifying the gitignore_parser
code, and changing the expectations of a gitignore_parser test.

For good measure, we add another gitignore_parser test and a few
dir_traversal tests to verify this new behavior.

This also fixes the fawltydeps:exclude_dir_overlap real_projects test,
which has been failing until now.