traverse_project: Warn when a given path conflicts with exclude pattern
Paths to be traversed (basepath/code/deps/pyenv) and exclude patterns
can be specified in various different places (config file, environment,
command line), and it is not always obvious to the used if/when they are
in direct conflict with each other. Consider for example:

  [tool.fawltydeps]
  code=["foo"]

in pyproject.toml, and

  --exclude=foo

or vice versa. The exclude pattern and the code path are in direct
conflict, and the code path will win (i.e. the exclude will be
disregarded). When this happens with paths and patterns that are given
by the user, we should produce a warning to help reduce user confusion.

Fix this by comparing given paths to exclude patterns in
traverse_project.find_sources(), and logging a warning message when a
conflict is detected (once per unique given path).

Add tests to verify the generation of appropriate warning messages.