traverse_project: Eliminate 'anchored patterns without base_dir' errors
Until now, we've refused "anchored" exclude patterns on the command-line
as we could not determine the appropriate base_dir for these patterns.

This turns out not to be very usable in practice: Most users are not
(and should not be) aware of why passing "foo/" on the command-line is
fine, but passing "foo/*" is not possible (and can only happen from a
.gitignore file after the future --exclude-from feature is merged).

We _should_ allow anchored patterns on the command line, and somehow
work around the fact that with --code/--deps/--pyenv we can end up with
very many paths indeed.

The implementation we settle on here and now is this: When an anchored
patterns is supplied, we will pass it to DirectoryTraversal.exclude()
_multiple_ times, once for each unique directory where we are searching
for code, deps, or pyenvs.

This should hopefully cause the least amount of surprise for our users:
Exclude patterns are applied and obeyed insofar as they do not
_directly_ conflict with a given path.

With this change, a couple of tests are removed, and their corresponding
behavior (aborting with "Anchored pattern without base_dir") is no
longer reachable from the traverse_project level or above.

A couple of other tests are adjusted to retain a more plausible scenario
(eliminating all code/deps/pyenvs from the given settings don't happen
in practice).

With this, we also fix the two remaining real_projects test cases:
fawltydeps:exclude_no_overlap and fawltydeps:exclude_anchored, and we're
finally back to an all-green test suite!