Allow tagging sample_projects/real_projects as POSIX- or Windows-only
Add the optional "posix_only" and "windows_only" boolean flags to the
TOML schema for test_sample_projects and test_real_projects. This flag
can be set either at the project level, or at the experiment level, and
will cause the associated experiments to be skipped on an unsupported
platform.