test_real_projects: Clean up how we create and populate venvs
This fixes issue #406.

Instead of generating lines of a shell script that need to be run with
shell=True in subprocess.run(), generate a sequence of argv-style
commands. These are safer (no shell involved) and more easily portable
to Windows (for one, our quoting of command line args on Windows was
dubious, at best...).

Also move the first and last commands in the previous shell script
(i.e. the preliminary removal of venv_dir, and the creation of the
venv_dir/.installed marker file) out of venv_script_lines() and into
the __call__() method.

Since these are no longer part of the generated commands, they are also
no longer part of the venv_hash(). If we make changes to these two steps
(without making any other changes that are captured by venv_hash(), we
will end up reusing a cache entry. If this is problematic, then some
other modification to change the venv_hash() must be added at the same
time.