tests.project_helpers: Improve encoding of Posix/Windows compatibility
Represent posix/windows compatibility as a proper enum.Flag that allows
test projects/experiments to declare their compatibility as either:

  - Compatibility.POSIX for Posix-only tests
  - Compatibility.WINDOWS for Windows-only tests
  - (Compatibility.POSIX | Compatibility.WINDOWS) for tests that can be
    run on either Posix or Windows (this is the default).

The compatibility of the current experiment is found by looking up the
.compatibility member of the experiment, or if unset, falling back to
the project's .compatibility member (which defaults to compatibility
with everything).

We then convert the current platform (sys.platform) into either
Compatibility.POSIX or Compatibility.WINDOWS, and check this against the
compatibility of the current experiment to see if if should be skipped
or not.

This allows adding further compatiblity flags in the future without
having to add more directives to the TOML schema for projects and
experiments.