Re-configure ruff to be much stricter
Enable ALL checks, then list + classify individual checks (or in some
cases, groups of checks) to be ignored. Most of these ignored checks
will be tackled and re-enabled in the next commits.

Finally, for individual instances of intentional check violations (or
false positives), add "noqa: ..." annotations to the code in question.
Some of these coincide with existing pylint annotations.