tests/project_helpers: Add LINUX and MACOS to Compatibility flags
These used to only be POSIX, but we soon need to differentiate between
them. POSIX remains as the OR-combination of these two.

Also extend the parser to throw a RuntimeError when a yet-unknown
platform is encountered. This is only in our _test_ code, so it serves
to flag a new/unexpected platform in our test suite without necessarily
causing FawltyDeps proper to crash on this new platform.