tests/real_projects/python-algorithms: Disable one test on MacOS
The all_reqs_installed test must be skipped on MacOS for the ~same
reasons as skipping it on Windows: The tensorflow package on (ARM-based)
MacOS is a placeholder package which causes a false-positive undeclared
dependency.

Newer versions of tensorflow (>=v2.15.1) seem to have stopped shipping
placeholder packages for ARM-based MacOS, so we might be able to revert
this some time in the future.