pyproject.toml: Add `uv` as an "extra" dependency for FawltyDeps
This allows users to install (or depend on) `"fawltydeps[uv]"` which
will then automatically bring in `uv` alongside `fawltydeps`.

This has no effect on our test/developer environments, as we install
`fawltydeps` without extras, but we still get `uv` via our dependency
on `nox[uv]`.

We should consider updating the FawltyDeps GitHub Action to install
`fawltydeps[uv]`, so that we can benefit from `uv` in that scenario as
well.