test_settings: Workaround issue in Pydantic v1.10.16
Pydantic v1.10.16 introduces the `pydantic.v1` namespace (adopted from
Pydantic v2) in order to make migrations from Pydantic v1 to v2 easier.

However, as detailed in issue #438 (and the corresponding upstream issue
https://github.com/pydantic/pydantic/issues/9656), in Pydantic v1.10.16
(when run on Python 3.12) objects imported via the `pydantic.v1`
namespace are not identical (in the `id(...)` sense) to the
corresponding objects imported via the regular `pydantic` namespace.

This causes problem when we want to catch exceptions raised by Pydantic:
the actual exception class being raised is not identical to the one we
`import` from `pydantic.v1.<something>`.

The workaround applied here is to ensure that we never import from the
`pydantic.v1` namespace when we're running on Pydantic v1.